hittarget: Use collider mesh instead of generated one.
