+++ VisualPinball.Unity/VisualPinball.Unity/VPT/HitTarget/DropTargetPackable.cs
@@ -31,7 +31,6 @@ namespace VisualPinball.Unity
 	public struct DropTargetColliderPackable
 	{
 		public bool IsMovable;
-		public bool IsLegacy;
 		public float Threshold;
 		public bool UseHitEvent;
 
@@ -39,7 +38,6 @@ namespace VisualPinball.Unity
 		{
 			return PackageApi.Packer.Pack(new DropTargetColliderPackable {
 				IsMovable = comp._isKinematic,
-				IsLegacy = comp.IsLegacy,
 				Threshold = comp.Threshold,
 				UseHitEvent = comp.UseHitEvent,
 			});
@@ -49,7 +47,6 @@ namespace VisualPinball.Unity
 		{
 			var data = PackageApi.Packer.Unpack<DropTargetColliderPackable>(bytes);
 			comp._isKinematic = data.IsMovable;
-			comp.IsLegacy = data.IsLegacy;
 			comp.Threshold = data.Threshold;
 			comp.UseHitEvent = data.UseHitEvent;
 		}
