+++ VisualPinball.Unity/VisualPinball.Unity/VPT/HitTarget/TargetColliderGenerator.cs
@@ -14,60 +14,44 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+using Unity.Collections;
 using Unity.Mathematics;
-using VisualPinball.Engine.VPT;
-using VisualPinball.Engine.VPT.HitTarget;
+using UnityEngine;
 
 namespace VisualPinball.Unity
 {
-	public abstract class TargetColliderGenerator
+	public class TargetColliderGenerator
 	{
-		protected readonly IApiColliderGenerator Api;
-		protected readonly TargetComponent Data;
-		protected readonly float4x4 Matrix;
+		private readonly Mesh _colliderMesh;
+		private readonly IApiColliderGenerator _api;
+		private readonly float4x4 _matrix;
 
-		protected TargetColliderGenerator(IApiColliderGenerator api, TargetComponent data, float4x4 matrix)
+		public TargetColliderGenerator(Mesh colliderMesh, IApiColliderGenerator api, float4x4 matrix)
 		{
-			Api = api;
-			Data = data;
-			Matrix = matrix;
+			_colliderMesh = colliderMesh;
+			_api = api;
+			_matrix = matrix;
 		}
 
-		// private protected void GenerateCollidables(Mesh hitMesh, ref EdgeSet addedEdges, bool setHitObject, ref ColliderReference colliders)  {
-		//
-		// 	// add the normal drop target as collidable but without hit event
-		// 	for (var i = 0; i < hitMesh.Indices.Length; i += 3) {
-		// 		var i0 = hitMesh.Indices[i];
-		// 		var i1 = hitMesh.Indices[i + 1];
-		// 		var i2 = hitMesh.Indices[i + 2];
-		//
-		// 		// NB: HitTriangle wants CCW vertices, but for rendering we have them in CW order
-		// 		var rgv0 = hitMesh.Vertices[i0].ToUnityFloat3();
-		// 		var rgv1 = hitMesh.Vertices[i1].ToUnityFloat3();
-		// 		var rgv2 = hitMesh.Vertices[i2].ToUnityFloat3();
-		//
-		// 		colliders.Add(new TriangleCollider(rgv0, rgv2, rgv1, GetColliderInfo(setHitObject)));
-		//
-		// 		if (addedEdges.ShouldAddHitEdge(i0, i1)) {
-		// 			colliders.Add(new Line3DCollider(rgv0, rgv2, GetColliderInfo(setHitObject)));
-		// 		}
-		// 		if (addedEdges.ShouldAddHitEdge(i1, i2)) {
-		// 			colliders.Add(new Line3DCollider(rgv2, rgv1, GetColliderInfo(setHitObject)));
-		// 		}
-		// 		if (addedEdges.ShouldAddHitEdge(i2, i0)) {
-		// 			colliders.Add(new Line3DCollider(rgv1, rgv0, GetColliderInfo(setHitObject)));
-		// 		}
-		// 	}
-		//
-		// 	// add collision vertices
-		// 	foreach (var vertex in hitMesh.Vertices) {
-		// 		colliders.Add(new PointCollider(vertex.ToUnityFloat3(), GetColliderInfo(setHitObject)));
-		// 	}
-		// }
+		internal void GenerateColliders(ref ColliderReference colliders)
+		{
+			using var meshDataArray = Mesh.AcquireReadOnlyMeshData(_colliderMesh);
+			var meshData = meshDataArray[0];
+			var subMesh = meshData.GetSubMesh(0); // todo loop through all sub meshes?
+			var unityVertices = new NativeArray<Vector3>(meshData.vertexCount, Allocator.TempJob);
+			var unityIndices = new NativeArray<int>(subMesh.indexCount, Allocator.TempJob);
+			meshData.GetVertices(unityVertices);
+			meshData.GetIndices(unityIndices, 0);
+
+			ColliderUtils.GenerateCollidersFromMesh(in unityVertices, in unityIndices, _matrix, _api.GetColliderInfo(), ref colliders);
+
+			unityVertices.Dispose();
+			unityIndices.Dispose();
+		}
 
 		protected ColliderInfo GetColliderInfo(bool setHitObject)
 		{
-			var info = Api.GetColliderInfo();
+			var info = _api.GetColliderInfo();
 			info.FireEvents = setHitObject && info.FireEvents;
 			return info;
 		}
